rename storage_mode -> storage
